# doc/api.yml
openapi: '3.0.2'
info:
  title: JCloudify Api
  version: 'latest'
servers:
  - url: https://jcloudify-api.preprod.hei.school
paths:
...
  /users/{userId}/applications/{applicationId}/environments:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/applicationId'
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Environment
      operationId: getApplicationEnvironments
      summary: Get environment list of an application
      description: list of environments, pagination was omitted because we might not have that many environments for a single app
      responses:
        '200':
          description: list of environments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentsResponse'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Environment
      operationId: crupdateApplicationEnvironments
      summary: Crupdate environments of an application.
      description: for each environment in the payload, it creates an environment if given id exists, updates it otherwise
      requestBody:
        description: >
          - All application relative properties such as SSM parameters are computed from the environment's application
          - Resource relative properties are linked to the plan of the Environment and can be set by using a different endpoint for plan upgrade with enum values such as 1024 or 2048 for memory, etc...
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateEnvironmentsRequestBody'
        required: true
      responses:
        '200':
          description: The crupdated environments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrupdateEnvironmentsResponse'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /users/{userId}/applications/{applicationId}/environments/{environmentId}/config:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/applicationId'
      - $ref: '#/components/parameters/environmentId'
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Environment
      operationId: getApplicationEnvironmentConfig
      summary: Get the config of an environment
      responses:
        '200':
          description: the config for the queried environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneOfPojaConf'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Environment
      operationId: configureApplicationEnv
      summary: configure an application's environment
      security:
        - bearerAuth: [ ]
      requestBody:
        description: the config to apply
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/OneOfPojaConf"
      responses:
        '200':
          description: the applied config
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/OneOfPojaConf"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    userId:
      name: userId
      in: path
      description: The identifier of the user
      schema:
        type: string
      required: true
    applicationId:
      name: applicationId
      in: path
      description: The identifier of the application
      schema:
        type: string
      required: true
    environmentId:
      name: environmentId
      in: path
      description: The identifier of the environment
      schema:
        type: string
      required: true
...
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    EnvironmentType:
      type: string
      enum:
        - PROD
        - PREPROD
    ##############################################################
    ###############  Poja Configurations #########################
    ##############################################################
    PojaVersionDiscriminator:
      description: will be the discriminator, possible values are the ones available as poja_version
      type: string
    PojaConf:
      type: object
      description: will serve to type all poja confs
      properties:
        version:
          $ref: "#/components/schemas/PojaVersionDiscriminator"
      discriminator:
        propertyName: 'version'
        mapping:
          '16.2.1': '#/components/schemas/PojaConfV16_2_1'
      required:
        - version

      ##############################################################
      ###############  Poja Conf V16.2.1 ###########################
      ##############################################################
    ConcurrencyConfV16_2_1:
      type: object
      properties:
        frontal_reserved_concurrent_executions_nb:
          type: string
        worker_reserved_concurrent_executions_nb:
          type: string
    ComputeConfV16_2_1:
      type: object
      properties:
        frontal_memory:
          type: number
          example: 2048
        frontal_function_timeout:
          type: number
          example: 30
        worker_memory:
          type: number
          example: 1024
        worker_batch:
          type: number
          example: 5
        worker_function_1_timeout:
          type: number
          example: 600
        worker_function_2_timeout:
          type: number
          example: 600
    AuroraConfV16_2_1:
      type: object
      properties:
        aurora_min_capacity:
          type: number
        aurora_max_capacity:
          type: number
        aurora_scale_point:
          type: number
        aurora_sleep:
          type: number
          example: 300
        aurora_auto_pause:
          type: boolean
          example: false
    ClientConfV16_2_1:
      type: object
      properties:
        ts_client_default_openapi_server_url:
          type: string
        ts_client_api_url_env_var_name:
          type: string
        aws_account_id:
          type: string
        codeartifact_repository_name:
          type: string
        codeartifact_domain_name:
          type: string
    DatabaseConfV16_2_1:
      type: object
      properties:
        database_non_root_username:
          type: string
        database_non_root_password:
          type: string
        prod_db_cluster_timeout:
          type: number
          example: 300
    CustomConfV16_2_1:
      type: object
      properties:
        custom_java_repositories:
          $ref: "#/components/schemas/StringMap"
        custom_java_deps:
          $ref: "#/components/schemas/StringMap"
        custom_java_env_vars:
          $ref: "#/components/schemas/StringMap"
    StringMap:
      type: object
      properties:
        additionalProperties:
          type: string
    PojaConfV16_2_1:
      type: object
      allOf:
        - $ref: "#/components/schemas/PojaConf"
      required:
        - version
      properties:
        version:
          $ref: "#/components/schemas/PojaVersionDiscriminator"
        with_snapstart:
          type: boolean
          example: false
        ses_source:
          type: string
          example: "noreply@nowhere.com"
        with_swagger_ui:
          type: boolean
          example: false
        package_full_name:
          type: string
          example: "school.hei"
        java_facade_it:
          type: string
          example: "FacadeIT"
        with_gen_clients:
          type: boolean
          example: false
        with_database:
          type: string
          example: "sqlite"
        jacoco_min_coverage:
          type: number
          example: 0.8
        with_publish_to_npm_registry:
          type: boolean
          example: false
        with_sentry:
          type: boolean
          example: false
        with_sonar:
          type: boolean
          example: false
        with_codeql:
          type: boolean
          example: false
        with_queue_nb:
          type: number
          enum:
            - 0
            - 1
            - 2
        with_file_storage:
          type: boolean
          example: true
        concurrency:
          $ref: '#/components/schemas/ConcurrencyConfV16_2_1'
        compute:
          $ref: '#/components/schemas/ComputeConfV16_2_1'
        aurora:
          $ref: '#/components/schemas/AuroraConfV16_2_1'
        database:
          $ref: '#/components/schemas/DatabaseConfV16_2_1'
        custom:
          $ref: '#/components/schemas/CustomConfV16_2_1'
        client:
          $ref: '#/components/schemas/ClientConfV16_2_1'
      ##############################################################
      ###############  Poja Conf V15.2.2 ###########################
      ##############################################################
    ##############################################################
    ###############  Poja Configurations #########################
    ##############################################################
    Environment:
      allOf:
        - $ref: '#/components/schemas/CrupdateEnvironment'
      type: object
      properties:
        state:
          type: string
          enum:
            - HEALTHY
            - UNHEALTHY
            - UNKNOWN
    CrupdateEnvironment:
      type: object
      properties:
        id:
          type: string
        environment_type:
          $ref: '#/components/schemas/EnvironmentType'
        archived:
          type: boolean
          default: false
...
    EnvironmentsResponse:
      type: object
      allOf:
        - properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Environment"
    CrupdateEnvironmentsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
    CrupdateEnvironmentsRequestBody:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CrupdateEnvironment"
...
    OneOfPojaConf:
      oneOf:
        - $ref: "#/components/schemas/PojaConfV16_2_1"
    #############################################
    ###############  OneOf and AnyOf Components ############
    #############################################

    #############################################
    ###############  Exceptions #################
    #############################################
    ExceptionModel:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/ExceptionModel'
      example:
        type: InternalServerException
        message: Unexpected error